Create NA answers for any not in spreadsheet.  Speed up import.
Any controls that we know to be AWWA that are not in the spreadsheet are built as answers in CSET with "NA".  
When looping through the spreadsheet rows, don't rely on what the spreadsheet says are the number of rows on the sheet -- stop iterating after we hit an empty row after processing the controls.