Fixed issues with adding and then deleting a contact
The API was not telling the UI what the key to the new record was, so a delete request was not possible.  Also fixed a problem with alert titles.  Also put email sending in a separate thread so that the UI doesn't have to wait while the SMTP server decided if it is going to throw an error or not.