Large refactor of configuration for maturity models.
Icon behavior is now defined in config.json by the maturity model name.  This was a perfect excuse to refactor the answer option properties, moving them into their corresponding behavior object.  
The model "lookup" was changed to use the model name instead of the ID, which makes it more intuitive to code to.