Large refactor of reference links to support URLs
We can now add rows to GEN_FILE that represent an external URL, rather than a local document.
Also, the URL for a local document now uses the 'referencedocument' API endpoint with the GEN_FILE file ID.  This is done for consistency and will also allow us to stage a Resource Library in the cloud in the future.  The CSET APi will be able to know if it should pull the requested document from the cloud.  The old URL that pulled directly from the local CSET API Documents folder would not support that.