Dropped REQUIREMENT_QUESTIONS table
The table was redundant with REQUIREMENT_QUESTIONS_SETS and was only being used in the requirement details view to show related questions.  We now use REQUIREMENT_QUESTIONS_SETS instead, and it provides the same mapping.  Dropping the table reduces work to keep the two tables in sync and removes a redundant table.