Componentized the warning display for IOD fields
Added the ability to set a display name on fields so that names with acronyms don't appear in mixed case, e.g., "IT ICS Staff Count".  More display names can be defined in the future, but for now, the fallback is to break the words on capital letters like before.