Refactor of CISA Assessor variable. Reports page tied to that now
There was a case where the Reports page disabled all of the report links and showed all of the missing fields when the assessment is in "normal" mode.  The reports page now looks at the assessorMode property of the assessment, rather than whether the user is a CISA assessor or not.