Upgrades assessment sectors from HSPD-7 to PPD-21
Addresses a requirement to align sector designations with PPD-21
standards. This change updates the sector ID in the assessment
demographics, removing subsector selections, and adds an acknowledgement
flag to inform the user of the change. It also cleans up PCII fields
from CIS demographics for JSON exports.

The upgrade logic is implemented to handle the conversion from old
HSPD-7 sector IDs to new PPD-21 sector IDs where applicable. A stored
procedure `update_demographic_sectors` is added to handle the update.