Improves logo color scheme for dark theme
Ensures the CSET logo adapts to both light and dark themes by modifying the logo component to accept a 'mode' input. This allows the logo to switch between a primary color fill and a white fill, improving visibility and aesthetics in the dark theme. The CSS class name has also been updated for clarity and consistency.

Fixes CSET-3360